rename keras1 to keras
