+++ pyspark/test/dev/run-tests
@@ -31,12 +31,12 @@ do
     $p -m pytest -v --doctest-modules ../../../pyspark/bigdl \
     --ignore=../../../pyspark/bigdl/dataset/ \
     --ignore=../../../pyspark/bigdl/util/tf_utils.py \
-    --ignore=../../../pyspark/bigdl/keras1 \
-    --ignore=../../../pyspark/test/bigdl/keras1/test_application.py \
+    --ignore=../../../pyspark/bigdl/keras \
+    --ignore=../../../pyspark/test/bigdl/keras/test_application.py \
     --ignore=../../../pyspark/bigdl/models/ && \
     $p -m pytest -v -n 4 ../../../pyspark/test/ \
     --ignore=../../../pyspark/test/bigdl/caffe/ \
-    --ignore=../../../pyspark/test/bigdl/keras1/ \
+    --ignore=../../../pyspark/test/bigdl/keras/ \
     --ignore=../../../pyspark/test/bigdl/test_utils.py
 
     exit_status=$?
